moved src/svg -> src/images
