+++ src/About.js
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 
 import { Footer, NavigationBar } from './components';
 import Card from './components/Card';
-import man from './svg/man.png';
+import man from './images/man.png';
 import H1 from './components/H1';
 
 export default function About() {
